requirements.txt
bokeh==2.0.2
boto3==1.13.6
botocore==1.16.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
html5lib==1.1
idna==2.9
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
numpy==1.18.3
packaging==20.3
pandas==1.0.3
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
text-unidecode==1.3
textile==4.0.1
tornado==6.0.4
typing-extensions==3.7.4.2
urllib3==1.25.8
webencodings==0.5.1
